fix: should not include unselected file when generate commit message #160
