+++ src/222/main/kotlin/cc/unitmesh/devti/vcs/VcsUtil.kt
@@ -1,13 +1,15 @@
 package cc.unitmesh.devti.vcs
 
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.components.service
+import com.intellij.openapi.vcs.VcsDataKeys
 import com.intellij.openapi.vcs.changes.Change
+import com.intellij.openapi.vcs.changes.CurrentContentRevision
+import com.intellij.vcs.commit.AbstractCommitWorkflowHandler
+import com.intellij.vcs.commit.CommitWorkflowUi
 
 object VcsUtil {
-    fun getChanges(e: AnActionEvent): List<Change>? {
-        val prompting = e.project?.service<VcsPrompting>() ?: return null
-        val changes = prompting.getChanges()
-        return changes.ifEmpty { null }
+    fun getCommitWorkFlowUi(e: AnActionEvent): CommitWorkflowUi? {
+        val commitWorkFlowHandler = e.getData(VcsDataKeys.COMMIT_WORKFLOW_HANDLER)
+        return (commitWorkFlowHandler as? AbstractCommitWorkflowHandler<*, *>)?.ui ?: return null
     }
 }
