+++ src/233/main/kotlin/cc/unitmesh/devti/vcs/VcsUtil.kt
@@ -6,20 +6,7 @@ import com.intellij.openapi.vcs.changes.Change
 import com.intellij.openapi.vcs.changes.CurrentContentRevision
 
 object VcsUtil {
-    fun getChanges(e: AnActionEvent): List<Change>? {
-        val commitWorkflowUi = e.getData(VcsDataKeys.COMMIT_WORKFLOW_UI) ?: return null
-
-        val changes = commitWorkflowUi.getIncludedChanges()
-        val unversionedFiles = commitWorkflowUi.getIncludedUnversionedFiles()
-
-        val unversionedFileChanges = unversionedFiles.map {
-            Change(null, CurrentContentRevision(it))
-        }
-
-        if (changes.isNotEmpty() || unversionedFileChanges.isNotEmpty()) {
-            return changes + unversionedFileChanges
-        }
-
-        return null
+    fun getCommitWorkFlowUi(e: AnActionEvent): CommitWorkflowUi? {
+        return e.getData(VcsDataKeys.COMMIT_WORKFLOW_UI) ?: return null
     }
 }
